feat(start.py): add start.py script to automate installation of dependencies and execution of different options
The start.py script is added to automate the installation of dependencies and provide a menu of options for executing different parts of the project. The script first installs the required dependencies by checking the operating system and using the appropriate package manager (brew, apt-get, or choco). Then, it presents a menu of options to the user to choose from. Each option corresponds to a different part of the project (01_Dataset_generation, 02_Model_Training, 03_Inference_Engine). The selected option is executed using the appropriate Python executable (python3 for Python 3.x and python for Python 2.x).